Scripts
This is the script that we have for player movement, this not working at the moment. We suspect that it is not since there is not a playable area on my machine. This playable area is being created by Cahill and will be uploaded shortly so more tests will come later.